add wind degrees to direction conversation

+++ WeatherApp/WeatherApp/Entites/GeographicWeather.swift
@@ -36,7 +36,7 @@ struct GeographicWeather: Codable {
 
     struct Wind: Codable {
         let speed: Double
-        let degrees: Int
+        let degrees: Double
         
         enum CodingKeys: String, CodingKey {
             case speed = "speed"
